Refactor testing; Use EC_KEY_generate_key
