+++ aws-lc-rs/tests/agreement_tests.rs
@@ -5,154 +5,7 @@
 
 extern crate alloc;
 
-use aws_lc_rs::{agreement, error, rand, test, test_file};
-
-#[test]
-fn agreement_traits() {
-    use alloc::vec::Vec;
-    use regex;
-    use regex::Regex;
-
-    let rng = rand::SystemRandom::new();
-    let private_key =
-        agreement::EphemeralPrivateKey::generate(&agreement::ECDH_P256, &rng).unwrap();
-
-    test::compile_time_assert_send::<agreement::EphemeralPrivateKey>();
-    //test::compile_time_assert_sync::<agreement::EphemeralPrivateKey>();
-
-    assert_eq!(
-        format!("{:?}", &private_key),
-        "EphemeralPrivateKey { algorithm: Algorithm { curve: P256 } }"
-    );
-
-    let public_key = private_key.compute_public_key().unwrap();
-    let pubkey_re = Regex::new(
-        "PublicKey \\{ algorithm: Algorithm \\{ curve: P256 \\}, bytes: \"[0-9a-f]+\" \\}",
-    )
-    .unwrap();
-    let pubkey_debug = format!("{:?}", &public_key);
-
-    assert!(
-        pubkey_re.is_match(&pubkey_debug),
-        "pubkey_debug: {pubkey_debug}"
-    );
-
-    #[allow(clippy::redundant_clone)]
-    let pubkey_clone = public_key.clone();
-    assert_eq!(public_key.as_ref(), pubkey_clone.as_ref());
-    assert_eq!(pubkey_debug, format!("{:?}", &pubkey_clone));
-
-    test::compile_time_assert_clone::<agreement::PublicKey>();
-    test::compile_time_assert_send::<agreement::PublicKey>();
-    //test::compile_time_assert_sync::<agreement::PublicKey>();
-
-    // Verify `PublicKey` implements `Debug`.
-    //
-    // TODO: Test the actual output.
-    let _: &dyn core::fmt::Debug = &public_key;
-
-    test::compile_time_assert_clone::<agreement::UnparsedPublicKey<&[u8]>>();
-    test::compile_time_assert_copy::<agreement::UnparsedPublicKey<&[u8]>>();
-    test::compile_time_assert_sync::<agreement::UnparsedPublicKey<&[u8]>>();
-
-    test::compile_time_assert_clone::<agreement::UnparsedPublicKey<Vec<u8>>>();
-    test::compile_time_assert_sync::<agreement::UnparsedPublicKey<Vec<u8>>>();
-
-    let bytes = [0x01, 0x02, 0x03];
-
-    let unparsed_public_key = agreement::UnparsedPublicKey::new(&agreement::X25519, &bytes);
-    let unparsed_pubkey_clone = unparsed_public_key;
-    assert_eq!(
-        format!("{unparsed_public_key:?}"),
-        r#"UnparsedPublicKey { algorithm: Algorithm { curve: Curve25519 }, bytes: "010203" }"#
-    );
-    assert_eq!(
-        format!("{unparsed_pubkey_clone:?}"),
-        r#"UnparsedPublicKey { algorithm: Algorithm { curve: Curve25519 }, bytes: "010203" }"#
-    );
-
-    let unparsed_public_key =
-        agreement::UnparsedPublicKey::new(&agreement::X25519, Vec::from(bytes));
-    #[allow(clippy::redundant_clone)]
-    let unparsed_pubkey_clone = unparsed_public_key.clone();
-    assert_eq!(
-        format!("{unparsed_public_key:?}"),
-        r#"UnparsedPublicKey { algorithm: Algorithm { curve: Curve25519 }, bytes: "010203" }"#
-    );
-    assert_eq!(
-        format!("{unparsed_pubkey_clone:?}"),
-        r#"UnparsedPublicKey { algorithm: Algorithm { curve: Curve25519 }, bytes: "010203" }"#
-    );
-}
-
-#[test]
-fn agreement_agree_ephemeral() {
-    let rng = rand::SystemRandom::new();
-
-    test::run(
-        test_file!("data/agreement_tests.txt"),
-        |section, test_case| {
-            assert_eq!(section, "");
-
-            let curve_name = test_case.consume_string("Curve");
-            let alg = alg_from_curve_name(&curve_name);
-            let peer_public =
-                agreement::UnparsedPublicKey::new(alg, test_case.consume_bytes("PeerQ"));
-
-            match test_case.consume_optional_string("Error") {
-                None => {
-                    let my_private_bytes = test_case.consume_bytes("D");
-                    let my_private = {
-                        let rng = test::rand::FixedSliceRandom {
-                            bytes: &my_private_bytes,
-                        };
-                        agreement::EphemeralPrivateKey::generate(alg, &rng)?
-                    };
-                    let my_public = test_case.consume_bytes("MyQ");
-                    let output = test_case.consume_bytes("Output");
-
-                    assert_eq!(my_private.algorithm(), alg);
-
-                    let computed_public = my_private.compute_public_key().unwrap();
-                    assert_eq!(computed_public.as_ref(), &my_public[..]);
-
-                    assert_eq!(my_private.algorithm(), alg);
-
-                    let result =
-                        agreement::agree_ephemeral(my_private, &peer_public, (), |key_material| {
-                            assert_eq!(key_material, &output[..]);
-                            Ok(())
-                        });
-                    assert_eq!(
-                        result,
-                        Ok(()),
-                        "Failed on private key: {:?}",
-                        test::to_hex(my_private_bytes)
-                    );
-                }
-
-                Some(_) => {
-                    fn kdf_not_called(_: &[u8]) -> Result<(), ()> {
-                        panic!(
-                            "The KDF was called during ECDH when the peer's \
-                         public key is invalid."
-                        );
-                    }
-                    let dummy_private_key = agreement::EphemeralPrivateKey::generate(alg, &rng)?;
-                    assert!(agreement::agree_ephemeral(
-                        dummy_private_key,
-                        &peer_public,
-                        (),
-                        kdf_not_called
-                    )
-                    .is_err());
-                }
-            }
-
-            Ok(())
-        },
-    );
-}
+use aws_lc_rs::{agreement, error, rand, test};
 
 #[test]
 fn test_agreement_ecdh_x25519_rfc_iterated() {
@@ -282,19 +135,5 @@ fn h(s: &str) -> Vec<u8> {
         Err(msg) => {
             panic!("{msg} in {s}");
         }
-    }
-}
-
-fn alg_from_curve_name(curve_name: &str) -> &'static agreement::Algorithm {
-    if curve_name == "P-256" {
-        &agreement::ECDH_P256
-    } else if curve_name == "P-384" {
-        &agreement::ECDH_P384
-    } else if curve_name == "P-521" {
-        &agreement::ECDH_P521
-    } else if curve_name == "X25519" {
-        &agreement::X25519
-    } else {
-        panic!("Unsupported curve: {curve_name}");
     }
 }
